CMake: Either build static or shared libraries
Always building shared and static libraries causes problems if the
platform we are compiling to does not support shared libraries.

This patch refactors the existing targets `zlib` and `zlibstatic` into a
single target `zlib` that by default will build the static version.

If a user requires the shared version is just a matter of setting the
CMake variable `BUILD_SHARED_LIBS` to On.